style: clean up storybook design system
